rename `gpt` related stuff to `llm`
